Only remove temporary dnadiff directory if it exists

+++ tests/generate_fixtures/generate_target_dnadiff_files.py
@@ -63,10 +63,10 @@ SHOW_COORDS_DIR = Path("../fixtures/dnadiff/targets/show_coords")
 # dnadiff generates many files; since we only need `.report` files,
 # we will save all files to a temporary location, move the report files
 # to a specific directory, and remove the rest.
-Path("./temp_dnadiff").mkdir(parents=True, exist_ok=True)
+Path("temp_dnadiff").mkdir(parents=True, exist_ok=True)
 
 DNADIFF_DIR = Path("../fixtures/dnadiff/targets/dnadiff_reports")
-DNADIFF_TEMP_DIR = Path("./temp_dnadiff")
+DNADIFF_TEMP_DIR = Path("temp_dnadiff")
 
 # Running comparisons
 comparisons = permutations([_.stem for _ in Path(INPUT_DIR).glob("*.f*")], 2)
@@ -144,4 +144,5 @@ for file in DNADIFF_TEMP_DIR.glob("*.report"):
     file.rename(DNADIFF_DIR / file.name)
 
 # remove dnadiff temporary directory
-shutil.rmtree(DNADIFF_TEMP_DIR)
+if DNADIFF_TEMP_DIR.exists() and DNADIFF_TEMP_DIR.is_dir():
+    shutil.rmtree(DNADIFF_TEMP_DIR)
